+++ my-app/package.json
@@ -13,6 +13,7 @@
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.0",
+    "router": "^1.3.5",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
